Move parameters into its own src file
