+++ src/app.py
@@ -1,3 +1,5 @@
+"""App."""
+
 import altair as alt
 import streamlit as st
