+++ src/penn_chime/defaults.py
@@ -6,9 +6,7 @@
 
 from collections import namedtuple
 
-# (0.02, 7) is 2%, 7 days
-# be sure to multiply by 100 when using as a default to the pct widgets!
-RateLos = namedtuple('RateLos', ('rate', 'length_of_stay'))
+from .utils import RateLos
 
 
 class Regions:
@@ -54,4 +52,3 @@ class Constants:
 
     def __repr__(self) -> str:
         return f"Constants(susceptible_default: {self.region.susceptible}, known_infected: {self.known_infected})"
-
