+++ src/penn_chime/utils.py
@@ -1,11 +1,25 @@
-from typing import Optional
+"""Utils."""
+
+from collections import namedtuple
 from datetime import datetime, timedelta
+from typing import Optional
+
 import numpy as np
 import pandas as pd
-import streamlit as st
 
-@st.cache
-def build_admissions_df(n_days, hosp, icu, vent) -> pd.DataFrame:
+
+# (0.02, 7) is 2%, 7 days
+# be sure to multiply by 100 when using as a default to the pct widgets!
+RateLos = namedtuple('RateLos', ('rate', 'length_of_stay'))
+
+
+def build_admissions_df(
+    n_days,
+    hosp,
+    icu,
+    vent,
+) -> pd.DataFrame:
+    """Build admis dataframe."""
     days = np.array(range(0, n_days + 1))
     data_dict = dict(zip(["day", "hosp", "icu", "vent"], [days, hosp, icu, vent]))
     projection = pd.DataFrame.from_dict(data_dict)
@@ -15,8 +29,13 @@ def build_admissions_df(n_days, hosp, icu, vent) -> pd.DataFrame:
     projection_admits["day"] = range(projection_admits.shape[0])
     return projection_admits
 
-@st.cache
-def build_census_df(projection_admits, hosp_los, icu_los, vent_los) -> pd.DataFrame:
+
+def build_census_df(
+    projection_admits,
+    hosp_los,
+    icu_los,
+    vent_los,
+) -> pd.DataFrame:
     """ALOS for each category of COVID-19 case (total guesses)"""
     n_days = np.shape(projection_admits)[0]
     los_dict = {
@@ -38,6 +57,7 @@ def build_census_df(projection_admits, hosp_los, icu_los, vent_los) -> pd.DataFr
     census_df = census_df[["day", "hosp", "icu", "vent"]]
     census_df = census_df.head(n_days)
     return census_df
+
 
 def add_date_column(
     df: pd.DataFrame,
