+++ src/test_app.py
@@ -1,3 +1,5 @@
+"""Tests."""
+
 import pytest
 import pandas as pd
